Improve RAPIDS support (require GPU, fix usage of GPU in parallel bagging)
