+++ tabular/src/autogluon/tabular/models/lr/lr_rapids_model.py
@@ -7,12 +7,13 @@ from autogluon.core.utils.try_import import try_import_rapids_cuml
 
 from .hyperparameters.parameters import get_param_baseline
 from .lr_model import LinearModel
+from .._utils.rapids_utils import RapidsModelMixin
 
 logger = logging.getLogger(__name__)
 
 
 # FIXME: If rapids is installed, normal CPU LinearModel crashes.
-class LinearRapidsModel(LinearModel):
+class LinearRapidsModel(RapidsModelMixin, LinearModel):
     """
     RAPIDS Linear model : https://rapids.ai/start.html
